Move cache to core module
